+++ betty/tests/data/aggregate/collection/test_keyed.py
@@ -1,7 +1,7 @@
 from collections.abc import Sequence
 
-from betty.collections import MutablePrimaryKeyCollection
-from betty.data.aggregate.collection.keyed import PrimaryKeyCollectionDefinition
+from betty.collection.keyed.adapter import MutableKeyedCollectionAdapter
+from betty.data.aggregate.collection.keyed import KeyedCollectionDefinition
 from betty.data.aggregate.record import FieldDefinition
 from betty.data.aggregate.record.mapping import TypedMappingDefinition
 from betty.data.indicator.selector import Key
@@ -10,7 +10,7 @@ from betty.portable import PortableData
 from betty.test_utils.locale.localizable import DUMMY_LOCALIZABLE
 
 
-class TestPrimaryKeyCollectionDefinition:
+class TestKeyedCollectionDefinition:
     _item = TypedMappingDefinition[dict[str, str]](
         cls=dict,
         label=DUMMY_LOCALIZABLE,
@@ -21,17 +21,18 @@ class TestPrimaryKeyCollectionDefinition:
             ),
         ],
     )
-    _sut_unordered = PrimaryKeyCollectionDefinition[dict[str, str]](
+    _sut_unordered = KeyedCollectionDefinition(
         value=_item,
         key=Key("key"),
-        ordered=False,
         label=DUMMY_LOCALIZABLE,
+        factory=lambda: MutableKeyedCollectionAdapter(key=lambda value: value["key"]),
     )
-    _sut_ordered = PrimaryKeyCollectionDefinition[dict[str, str]](
+    _sut_ordered = KeyedCollectionDefinition(
         value=_item,
         key=Key("key"),
-        ordered=True,
+        order_dump=True,
         label=DUMMY_LOCALIZABLE,
+        factory=lambda: MutableKeyedCollectionAdapter(key=lambda value: value["key"]),
     )
     _portable_unordered: PortableData = {
         "my_first_key": {
@@ -53,20 +54,24 @@ class TestPrimaryKeyCollectionDefinition:
 
     def test_load__unordered(self) -> None:
         data = self._sut_unordered.porter.load(self._portable_unordered)
-        assert isinstance(data, MutablePrimaryKeyCollection)
+        assert isinstance(data, MutableKeyedCollectionAdapter)
         assert data["my_first_key"]["key"] == "my_first_key"
         assert data["my_first_key"]["other_element"] == "my_first_other_element"
 
     def test_load__ordered(self) -> None:
         data = self._sut_ordered.porter.load(self._portable_ordered)
-        assert isinstance(data, MutablePrimaryKeyCollection)
+        assert isinstance(data, MutableKeyedCollectionAdapter)
         assert data["my_first_key"]["key"] == "my_first_key"
         assert data["my_first_key"]["other_element"] == "my_first_other_element"
 
     def test_dump__unordered(self) -> None:
-        data = MutablePrimaryKeyCollection(self._values, key=lambda value: value["key"])
+        data = MutableKeyedCollectionAdapter(
+            self._values, key=lambda value: value["key"]
+        )
         assert self._sut_unordered.porter.dump(data) == self._portable_unordered
 
     def test_dump__ordered(self) -> None:
-        data = MutablePrimaryKeyCollection(self._values, key=lambda value: value["key"])
+        data = MutableKeyedCollectionAdapter(
+            self._values, key=lambda value: value["key"]
+        )
         assert self._sut_ordered.porter.dump(data) == self._portable_ordered
